Split db/operations.ts into 8 domain modules
operations.ts (2,699 LOC) was a single monolithic file
mixing entity upserts, queries, sync management, user
settings, profiles, analytics, and backup logic.

Split into:
- operations/types.ts     shared types and helpers
- operations/entities.ts  entity upsert functions
- operations/queries.ts   fetch/delete query functions
- operations/sync.ts      sync state and run management
- operations/profiles.ts  user and repository profiles
- operations/user-settings.ts  preferences and holidays
- operations/analytics.ts dashboard stats and counts
- operations/backup.ts    DB backup record management

operations.ts is now a barrel re-exporting all modules.
All existing import paths remain unchanged.

Closes #369
